+++ bionty/models.py
@@ -250,6 +250,8 @@ class BioRecord(Record, HasParents, CanCurate):
     def import_source(
         cls,
         source: Source | None = None,
+        update_records: bool = False,
+        *,
         organism: str | Record | None = None,
         ontology_ids: list[str] | None = None,
         # field: FieldAttr | None = None,
@@ -261,6 +263,7 @@ class BioRecord(Record, HasParents, CanCurate):
 
         Args:
             source: Source record to import records from.
+            update_records: Whether to update existing records with the new source.
             organism: Organism name or record.
             ontology_ids: List of ontology ids to save. Default is None (save all).
             ignore_conflicts: Whether to ignore conflicts during bulk record creation.
@@ -269,17 +272,26 @@ class BioRecord(Record, HasParents, CanCurate):
 
             import bionty as bt
 
-            bt.CellType.import_source()
-        """
-        from .core._add_ontology import add_ontology_from_df
+            # import all records from a source
+            bt.CellType.import_source(source1)
 
-        add_ontology_from_df(
-            registry=cls,
-            ontology_ids=ontology_ids,
-            organism=organism,
-            source=source,
-            ignore_conflicts=ignore_conflicts,
-        )
+            # update existing records with a new source
+            bt.CellType.import_source(source2, update_records=True)
+        """
+        if update_records:
+            from .core._source import update_records_to_source
+
+            update_records_to_source(cls, source)
+        else:
+            from .core._add_ontology import add_ontology_from_df
+
+            add_ontology_from_df(
+                registry=cls,
+                ontology_ids=ontology_ids,
+                organism=organism,
+                source=source,
+                ignore_conflicts=ignore_conflicts,
+            )
 
     @classmethod
     def add_source(
@@ -372,121 +384,6 @@ class BioRecord(Record, HasParents, CanCurate):
         logger.important("source added!")
 
         return new_source
-
-    @classmethod
-    def upgrade_source(cls, source: Source) -> None:
-        """Upgrade the existing records associated with old source to the new source.
-
-        Args:
-            source: Source record to update from another entity.
-
-        Returns:
-            A Source record with this entity.
-
-        Example::
-
-            import bionty as bt
-
-        """
-        import lamindb as ln
-
-        from ._source import filter_public_df_columns
-
-        entity = cls.__get_name_with_module__()
-
-        if entity != source.entity:
-            raise ValueError(
-                f"please pass a source record of the same entity: {entity}"
-            )
-
-        # find records that need to be upgraded
-        filter_kwargs = {
-            "source__entity": entity,
-            "source__name": source.name,
-        }
-        if hasattr(cls, "organism_id"):
-            filter_kwargs["organism__name"] = source.organism
-        records = cls.filter(**filter_kwargs).exclude(source=source).all()
-        if len(records) == 0:
-            return
-
-        # determine the ontology ID field to use
-        ontology_id_field = getattr(cls, "_ontology_id_field", "ontology_id")
-        name_field = getattr(cls, "_name_field", "name")
-
-        # get the new data from the source
-        public_df = filter_public_df_columns(cls, cls.public(source=source))
-        public_df.rename(columns={"parents": "_parents"}, inplace=True)
-        if ontology_id_field not in public_df.columns:
-            ontology_id_field = (
-                "stable_id" if "stable_id" in public_df.columns else None
-            )
-            if not ontology_id_field:
-                raise ValueError(
-                    f"'{ontology_id_field}' column is not found in the source dataframe."
-                )
-
-        # a dictionary of records indexed by ontology ID
-        public_dict = {row[ontology_id_field]: row for _, row in public_df.iterrows()}
-
-        # filter records that have matching ontology IDs
-        records = records.filter(
-            **{ontology_id_field + "__in": set(public_dict.keys())}
-        ).all()
-
-        # update records without artifacts
-        # simply update the record if no artifacts are associated
-        records_without_artifacts = records.filter(artifacts=None).all()
-        records_to_update = []
-        if records_without_artifacts:
-            for record in records_without_artifacts:
-                ontology_id = getattr(record, ontology_id_field)
-                # update all fields from the new source
-                for col in public_df.columns:
-                    setattr(record, col, public_dict[ontology_id].get(col))
-                record.source_id = source.id
-                records_to_update.append(record)
-
-        # handle records with artifacts
-        records_with_artifacts = records.exclude(artifacts=None).all()
-        ontology_ids_for_new_records = []
-        for record in records_with_artifacts:
-            ontology_id = getattr(record, ontology_id_field)
-
-            # if name changed, create a new record
-            if (
-                ontology_id in public_dict
-                and hasattr(record, name_field)
-                and getattr(record, name_field)
-                != public_dict[ontology_id].get(name_field)
-            ):
-                ontology_ids_for_new_records.append(ontology_id)
-            # otherwise, update the existing record
-            else:
-                for col in public_df.columns:
-                    setattr(record, col, public_dict[ontology_id].get(col))
-                record.source_id = source.id
-                records_to_update.append(record)
-        # bulk update records with artifacts where name didn't change
-        if records_to_update:
-            logger.info(f"updating {len(records_to_update)} records...")
-            ln.save(records_to_update)
-            logger.success(f"{len(records_to_update)} records updated!")
-
-        # create new records for those with name changes
-        if ontology_ids_for_new_records:
-            logger.info(f"creating {len(ontology_ids_for_new_records)} new records...")
-            cls.from_values(
-                ontology_ids_for_new_records,
-                field=getattr(cls, ontology_id_field),
-                source=source,
-            ).save()
-            logger.success(f"{len(ontology_ids_for_new_records)} new records created!")
-
-        # set the source as currently used
-        if not source.currently_used:
-            source.currently_used = True
-            source.save()
 
     @classmethod
     def public(
